Add files via upload
I have created a Jupyter Notebook that details my clustering analysis on a country dataset. Using Python and popular libraries such as Scikit-learn, Pandas, and NumPy, I applied various clustering algorithms to group countries based on their similarities in demographics, economics, and other factors. The notebook includes detailed documentation on the methods used and the results obtained, as well as visualizations that help to better understand the clusters. You can easily reproduce the analysis and explore the code yourself by downloading and running the notebook.